+++ src/Inputs/Multiselect.js
@@ -4,7 +4,7 @@ import {useEffect, useRef, useState, useCallback} from 'react'
 import PropTypes from 'prop-types'
 import ReactSelect from 'react-select'
 import Creatable from 'react-select/creatable'
-import {isMobile} from '../utils'
+import {isMobile, convertDelimitedValueIntoLabelValueArray, convertLabelValueArrayIntoDelimitedValue} from '../utils'
 import ValidationErrorIcon from '../ValidationErrorIcon'
 import useTheme from '../theme/useTheme'
 
@@ -27,7 +27,11 @@ const Multiselect = props => {
     onChange,
     autoComplete,
     interactive = true,
-    style = {}
+    style = {},
+    delimit,
+    delimiter = '¤',
+    stringify,
+    isClearable = true
   } = props
 
   const {
@@ -53,7 +57,7 @@ const Multiselect = props => {
 
   const [input, changeInput] = useState({Select: !interactive ? Creatable : allowcreate ? Creatable : ReactSelect})
   const [isRequiredFlag, updateIsRequiredFlag] = useState(required && requiredWarning && !value.length)
-  const [menuIsOpen, updateIsMenuOpen] = useState(false)
+  const [menuIsOpen, updateIsMenuOpen] = useState({})
   const [menuPlacement, updateMenuPlacement] = useState('bottom')
   const [fieldPosition, updateFieldPosition] = useState(0)
   const [selectValue, updateSelectValue] = useState([])
@@ -63,10 +67,10 @@ const Multiselect = props => {
   const inputContainer = useRef(null)
 
   const openMenu = useCallback(() => {
-    if (!readonly && !disabled && !menuIsOpen) {
-      updateIsMenuOpen(true)
+    if (!readonly && !disabled && !menuIsOpen[name]) {
+      updateIsMenuOpen({...menuIsOpen, [name]: true})
     }
-  }, [readonly, disabled, updateIsMenuOpen, menuIsOpen])
+  }, [readonly, disabled, updateIsMenuOpen, menuIsOpen, name])
 
   const setMenuOpenPosition = useCallback(() => {
     const placement = fieldPosition < (viewPortHeight / 2) ? 'bottom' : 'top'
@@ -74,14 +78,16 @@ const Multiselect = props => {
   }, [fieldPosition, updateMenuPlacement])
 
   const handleInputBlur = useCallback(() => {
-    menuIsOpen && updateIsMenuOpen(false)
+    menuIsOpen[name] && updateIsMenuOpen({...menuIsOpen, [name]: false})
     setIsFocused(false)
-  }, [menuIsOpen, updateIsMenuOpen])
+  }, [menuIsOpen, updateIsMenuOpen, name])
 
   const setInputFieldPosition = useCallback(() => {
-    const position = inputContainer.current.getBoundingClientRect().top
-    if (fieldPosition !== position) {
-      updateFieldPosition(position)
+    if (inputContainer.current) {
+      const position = inputContainer.current.getBoundingClientRect().top
+      if (fieldPosition !== position) {
+        updateFieldPosition(position)
+      }
     }
     setTimeout(openMenu) // this needs to be refactored so it actually updates with react instead of hacking around the problem - JRA 12/18/2019
   }, [openMenu, fieldPosition])
@@ -97,10 +103,18 @@ const Multiselect = props => {
     setIsFocused(true)
   }, [handleInputClick])
 
+  const closeMenuOnScroll = useCallback(e => {
+    let menuOpenState = false
+    if (e && e.target && e.target.classList) {
+      menuOpenState = e.target.classList.contains('gfb-input__menu-list') && menuIsOpen[name]
+    }
+    updateIsMenuOpen({...menuIsOpen, [name]: menuOpenState})
+  }, [menuIsOpen, name, updateIsMenuOpen])
+
   useEffect(() => {
     let formattedOptions = keyword.options || []
     if (!formattedOptions) formattedOptions = []
-    if (typeof formattedOptions === 'string') formattedOptions = formattedOptions.split('¤')
+    if (typeof formattedOptions === 'string') formattedOptions = formattedOptions.split(delimiter)
     if (formattedOptions.toJS) formattedOptions = formattedOptions.toJS()
 
     const duplicate = {}
@@ -123,7 +137,7 @@ const Multiselect = props => {
     })
 
     updateSelectOptions(formattedOptions)
-  }, [keyword.options])
+  }, [delimiter, keyword.options])
 
   useEffect(() => {
     setMenuOpenPosition()
@@ -138,58 +152,23 @@ const Multiselect = props => {
   }, [updateIsRequiredFlag, required, requiredWarning, value])
 
   useEffect(() => {
-    let formattedValue = value
-    // first lets try to get this value normalized to what react-select wants, which is an array of values
-    if (!formattedValue) formattedValue = []
-    if (formattedValue.toJS) formattedValue = formattedValue.toJS()
-    if (typeof formattedValue === 'string') formattedValue = formattedValue.split('¤')
-    if (!Array.isArray(formattedValue) && typeof formattedValue === 'object') {
-      formattedValue = Object.values(formattedValue)
-    }
-    if (!Array.isArray(formattedValue)) {
-      console.warn('The field', name, 'is a multiselect but its value was not a valid multi value. Multivalues should be a delimited string or an array of values, but instead got', value) //eslint-disable-line
-      formattedValue = []
-    }
+    updateSelectValue(convertDelimitedValueIntoLabelValueArray({value, delimit, delimiter, options}))
+  }, [value, updateSelectValue, name, delimit, delimiter, stringify, options])
 
-    const duplicate = {}
-    // lets filter out any blanks they may have snuck in
-    formattedValue.filter(value => {
-      if (typeof value === 'object') value = value.value // if value is an object but does not have a value key, we are going to drop the value as well - JRA 12/19/2019
-      if (!value) return false
-      if (!duplicate[value]) {
-        duplicate[value] = true
-        return true
-      }
-    })
-
-    // now lets make sure each value in the array is a {label, value} object
-    formattedValue = formattedValue.map(value => {
-      if (typeof value === 'string') {
-        value = {label: value, value}
-      }
-      if (typeof value === 'object' && !value.label) {
-        value.label = value.value
-      }
-      return value
-    })
-
-    updateSelectValue(formattedValue)
-  }, [value, updateSelectValue, name])
-
-  const handleOnKeyDown = useCallback(() => {
-    if (!menuIsOpen) openMenu()
-    onKeyDown()
-  }, [onKeyDown, menuIsOpen, openMenu])
-
-  const handleChange = useCallback(e => {
+  const handleChange = useCallback(val => {
     onChange({
       target: {
         name,
-        value: e === null ? [] : e
+        value: convertLabelValueArrayIntoDelimitedValue({value: val, delimiter, delimit, stringify})
       }
     })
-    menuIsOpen && updateIsMenuOpen(false)
-  }, [onChange, name, menuIsOpen])
+    menuIsOpen[name] && updateIsMenuOpen({...menuIsOpen, [name]: false})
+  }, [onChange, name, delimiter, delimit, stringify, menuIsOpen])
+
+  const handleOnKeyDown = useCallback(() => {
+    if (!menuIsOpen[name]) openMenu()
+    onKeyDown()
+  }, [onKeyDown, menuIsOpen, openMenu, name])
 
   const {Select} = input
 
@@ -220,11 +199,11 @@ const Multiselect = props => {
         className={className}
         classNamePrefix='gfb-input'
         tabIndex={tabIndex}
-        autofocus={autofocus}
-        isClearable
+        autoFocus={autofocus}
+        closeMenuOnScroll={!isMobile ? closeMenuOnScroll : undefined}
+        isClearable={isClearable}
         isDisabled={disabled || readonly || !interactive}
         menuPortalTarget={document.body}
-        menuShouldBlockScroll
         isMulti
         name={name}
         options={options}
@@ -232,7 +211,7 @@ const Multiselect = props => {
         onFocus={handleOnFocus}
         onKeyDown={handleOnKeyDown}
         onBlur={handleInputBlur}
-        menuIsOpen={!isMobile ? menuIsOpen : undefined}
+        menuIsOpen={!isMobile ? menuIsOpen[name] : undefined}
         menuPlacement={!isMobile ? menuPlacement : undefined}
         value={selectValue}
         defaultValue={selectValue}
@@ -295,5 +274,9 @@ Multiselect.propTypes = {
   onKeyDown: PropTypes.func,
   autoComplete: PropTypes.string,
   interactive: PropTypes.bool,
-  style: PropTypes.object
+  style: PropTypes.object,
+  stringify: PropTypes.bool,
+  delimiter: PropTypes.string,
+  delimit: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
+  isClearable: PropTypes.bool
 }